it wasn't so shrimple after all...
If the offset was larger than the beat crochet, it'd start overwriting itself before it could get a chance to do its thing.

Now we're using an array here -- it's really ugly... Could have it as a singular looping timer on its own, but then again I'd still need a similar array thingymagig to pass the event handling thingy for stuff like propagation cancelling. Which, um. Uhh!